backend/requirements.txt
fastapi==0.115.12
pydantic-settings==2.8.1
PyJWT==2.9.0
python-multipart==0.0.20
setuptools<81
uvicorn[standard]==0.34.0
